+++ lib/presentation/onboarding/screens/onboarding_screen.dart
@@ -6,7 +6,7 @@ import 'package:on_time_front/core/di/di_setup.dart';
 import 'package:on_time_front/presentation/onboarding/components/preparation_reordarable_list.dart';
 import 'package:on_time_front/presentation/onboarding/preparation_name_select/cubit/preparation_name/preparation_name_cubit.dart';
 import 'package:on_time_front/presentation/onboarding/preparation_name_select/screens/preparation_select_list.dart';
-import 'package:on_time_front/presentation/onboarding/components/preparation_time_input_list.dart';
+import 'package:on_time_front/presentation/onboarding/preparation_time/screens/preparation_time_input_list.dart';
 import 'package:on_time_front/presentation/onboarding/components/schedule_spare_time_input.dart';
 import 'package:on_time_front/presentation/onboarding/cubit/onboarding/onboarding_cubit.dart';
 import 'package:on_time_front/presentation/onboarding/preparation_order/cubit/preparation_order_cubit.dart';
